BF: Reverts all commits from #2278 and #2279
The original issue from Discourse appeared to be a problem with interpretting
mixed EOLs in Coder and code components. The original fix
attempted to convert EOLs to those defined in the preferences to provide
uniform EOLS. However, this did not address the problem described in
the original post.
